// BackOffice/src/app/model/product-view-model.model.ts
export class ProductViewModel {
    Id:number;
    Name: string;
    DateCreated: Date;
    Description: string;
    Details: string;
    LanguageId: string;
    OriginalPrice: number;
    Price: number;
    SeoAlias: string;
    SeoDescription: string;
    SeoTitle: string;
    Stock: number;
    ViewCount: number;
    Rate: number;
}